requirements.txt
# === Core Backend ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4

# === AI / ML Models ===
open-clip-torch==2.29.0
sentence-transformers==3.3.1
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0

# === OCR & Document Processing ===
easyocr==1.7.2
PyMuPDF==1.25.3
python-docx==1.1.2

# === Vector Database ===
chromadb==0.6.3

# === File Monitoring ===
watchdog==6.0.0

# === Google Gemini ===
google-generativeai==0.8.4

# === Image Metadata ===
exifread==3.0.0

# === Frontend ===
streamlit==1.41.1
requests==2.32.3
plotly==5.24.1

# === Utilities ===
numpy>=1.24.0
